+++ ass_man/views.py
@@ -6,10 +6,12 @@ from rest_framework.response import Response
 
 # API
 from rest_framework import viewsets
-from ass_man.serializers import UserSerializer, InstanceShortSerializer, InstanceSerializer, \
-    ModelShortSerializer, ModelSerializer,  RackSerializer
+from ass_man.serializers import (InstanceShortSerializer,
+                                 InstanceSerializer,
+                                 ModelShortSerializer,
+                                 ModelSerializer,
+                                 RackSerializer)
 # Auth
-from django.contrib.auth.models import User
 from rest_framework.permissions import IsAuthenticated, IsAdminUser, AllowAny
 # Project
 from ass_man.models import Model, Instance, Rack
@@ -18,20 +20,6 @@ ADMIN_ACTIONS = {'create', 'update', 'partial_update', 'destroy'}
 
 
 # Docs for ModelViewSet: https://www.django-rest-framework.org/api-guide/viewsets/#modelviewset
-class UserViewSet(viewsets.ModelViewSet):
-    # API endpoint that allows users to be viewed or edited.
-
-    # def get_permissions(self):
-    #     # Instantiates and returns the list of permissions that this view requires.
-    #     if self.action in ADMIN_ACTIONS:
-    #         permission_classes = [IsAdminUser]
-    #     else:
-    #         permission_classes = [IsAuthenticated]
-    #     return [permission() for permission in permission_classes]
-
-    permission_classes = [AllowAny]
-    queryset = User.objects.all().order_by('-date_joined')
-    serializer_class = UserSerializer
 
 
 class ModelViewSet(viewsets.ModelViewSet):
