+++ frontend/frontend-react/.gitignore
@@ -10,6 +10,7 @@
 
 # production
 /build
+build/
 
 # misc
 .DS_Store
