+++ requirements.txt
@@ -1,14 +1,20 @@
 asgiref==3.2.3
 astroid==2.3.3
+cffi==1.13.2
+cryptography==2.8
 dj-database-url==0.5.0
 Django==3.0.2
 django-cors-headers==3.2.1
+django-rest-knox==4.1.0
 djangorestframework==3.11.0
+djangorestframework-simplejwt==4.4.0
 gunicorn==20.0.4
 isort==4.3.21
 lazy-object-proxy==1.4.3
 mccabe==0.6.1
 psycopg2-binary==2.8.4
+pycparser==2.19
+PyJWT==1.7.1
 pylint==2.4.4
 python-dotenv==0.10.5
 pytz==2019.3
